+++ src/Components/PaypalExpress/Route/FinishCheckoutRoute.php
@@ -6,11 +6,10 @@ namespace Kiener\MolliePayments\Components\PaypalExpress\Route;
 use Kiener\MolliePayments\Components\PaypalExpress\PayPalExpress;
 use Kiener\MolliePayments\Components\PaypalExpress\PaypalExpressException;
 use Kiener\MolliePayments\Service\CartServiceInterface;
-use Kiener\MolliePayments\Service\CustomFieldsInterface;
 use Kiener\MolliePayments\Service\SettingsService;
 use Kiener\MolliePayments\Struct\Address\AddressStruct;
+use Mollie\Shopware\Entity\Cart\MollieShopwareCart;
 use Shopware\Core\Framework\Plugin\Exception\DecorationPatternException;
-use Shopware\Core\Framework\Struct\ArrayStruct;
 use Shopware\Core\System\SalesChannel\SalesChannelContext;
 
 /**
@@ -47,20 +46,12 @@ class FinishCheckoutRoute extends AbstractFinishCheckoutRoute
         }
 
         $cart = $this->cartService->getCalculatedMainCart($context);
-        $cartExtension = $cart->getExtension(CustomFieldsInterface::MOLLIE_KEY);
+        $mollieShopwareCart = new MollieShopwareCart($cart);
 
-        $payPalExpressSessionId = null;
-        $acceptedDataProtection = null;
+        $payPalExpressSessionId = $mollieShopwareCart->getPayPalExpressSessionID();
+        $acceptedDataProtection = $mollieShopwareCart->isDataProtectionAccepted();
 
-        if ($cartExtension instanceof ArrayStruct) {
-            $payPalExpressSessionId = $cartExtension[CustomFieldsInterface::PAYPAL_EXPRESS_SESSION_ID_KEY] ?? null;
-            if ($settings->isRequireDataProtectionCheckbox()) {
-                $acceptedDataProtection = $cartExtension[CustomFieldsInterface::ACCEPTED_DATA_PROTECTION] ?? 0;
-            }
-        }
-
-
-        if ($payPalExpressSessionId === null) {
+        if ($payPalExpressSessionId === '') {
             throw PaypalExpressException::cartSessionIdIsEmpty();
         }
 
@@ -70,23 +61,23 @@ class FinishCheckoutRoute extends AbstractFinishCheckoutRoute
         $methodDetails = $payPalExpressSession->methodDetails;
 
 
-        if (! property_exists($methodDetails, 'shippingAddress') || $methodDetails->shippingAddress === null) {
+        if (!property_exists($methodDetails, 'shippingAddress') || $methodDetails->shippingAddress === null) {
             throw PaypalExpressException::shippingAddressMissing();
         }
-        if (! property_exists($methodDetails, 'billingAddress') || $methodDetails->billingAddress === null) {
+        if (!property_exists($methodDetails, 'billingAddress') || $methodDetails->billingAddress === null) {
             throw PaypalExpressException::billingAddressMissing();
         }
 
         $billingAddress = null;
 
         $mollieShippingAddress = $methodDetails->shippingAddress;
-        if (! property_exists($mollieShippingAddress, 'phone')) {
+        if (!property_exists($mollieShippingAddress, 'phone')) {
             $mollieShippingAddress->phone = '';
         }
-        if (! property_exists($mollieShippingAddress, 'streetAdditional')) {
+        if (!property_exists($mollieShippingAddress, 'streetAdditional')) {
             $mollieShippingAddress->streetAdditional = '';
         }
-        if (! property_exists($mollieShippingAddress, 'email')) {
+        if (!property_exists($mollieShippingAddress, 'email')) {
             $mollieShippingAddress->email = '';
         }
 
@@ -123,15 +114,12 @@ class FinishCheckoutRoute extends AbstractFinishCheckoutRoute
         $this->paypalExpress->prepareCustomer($shippingAddress, $context, $acceptedDataProtection, $billingAddress);
 
         # we have to update the cart extension before a new user is created and logged in, otherwise the extension is not saved
-        $cartExtension = new ArrayStruct([
-            CustomFieldsInterface::PAYPAL_EXPRESS_AUTHENTICATE_ID => $payPalExpressSession->authenticationId
-        ]);
-        $cart->addExtension(CustomFieldsInterface::MOLLIE_KEY, $cartExtension);
+        $mollieShopwareCart = new MollieShopwareCart($cart);
+        $mollieShopwareCart->setPayPalExpressAuthenticateId($payPalExpressSession->authenticationId);
 
+        $cart = $mollieShopwareCart->getCart();
         $this->cartService->updateCart($cart);
-
         $this->cartService->persistCart($cart, $context);
-
 
         return new FinishCheckoutResponse($payPalExpressSession->id, $payPalExpressSession->authenticationId);
     }