+++ src/Resources/app/storefront/src/mollie-payments/services/ExpressAddToCart.js
@@ -1,25 +1,37 @@
 import BuyButtonRepository from '../repository/BuyButtonRepository';
 
 export default class ExpressAddToCart {
-    addItemToCart(button) {
-        const buyButtonRepository = new BuyButtonRepository();
-        const buyButton = buyButtonRepository.find(button);
-        if (!(buyButton instanceof HTMLButtonElement)) {
-            return;
 
-        }
-        const buyButtonForm = buyButton.closest('form');
-        if (!(buyButtonForm instanceof HTMLFormElement)) {
-            return;
-        }
+    addItemToCartOrSkip(button) {
 
-        const formData = new FormData(buyButtonForm);
-        formData.delete('redirectTo');
-        formData.append('isExpressCheckout', '1');
+        return new Promise((resolve, reject) => {
 
-        fetch(buyButtonForm.action, {
-            method: buyButtonForm.method,
-            body: formData,
-        }).finally(() =>{});
+            const buyButtonRepository = new BuyButtonRepository();
+            const buyButton = buyButtonRepository.find(button);
+
+            if (!(buyButton instanceof HTMLButtonElement)) {
+                resolve();
+            }
+
+            const buyButtonForm = buyButton.closest('form');
+
+            if (!(buyButtonForm instanceof HTMLFormElement)) {
+                resolve();
+            }
+
+            const formData = new FormData(buyButtonForm);
+            formData.delete('redirectTo');
+            formData.append('isExpressCheckout', '1');
+
+            fetch(buyButtonForm.action, {
+                method: buyButtonForm.method,
+                body: formData,
+            })
+                .then(() => {
+                    resolve()
+                })
+                .catch(error => reject(error));
+        });
     }
+
 }