+++ src/Service/CustomFieldsInterface.php
@@ -28,8 +28,9 @@ interface CustomFieldsInterface
 
     public const THIRD_PARTY_PAYMENT_KEY = 'third_party_payment_id';
 
-    public const PAYPAL_EXPRESS_SESSION_ID_KEY = 'mollie_ppe_session_id';
+    public const SINGLE_PRODUCT_EXPRESS_CHECKOUT = 'mollie_single_product_express_checkout';
 
+    public const PAYPAL_EXPRESS_SESSION_ID_KEY = 'mollie_ppe_session_id';
     public const PAYPAL_EXPRESS_AUTHENTICATE_ID = 'mollie_ppe_auth_id';
 
     public const ACCEPTED_DATA_PROTECTION = 'acceptedDataProtection';